Move test folder out of package module
