+++ test/test_verification_widget.py
